+++ src/dockerFolding.ts
@@ -4,12 +4,12 @@
 * ------------------------------------------------------------------------------------------ */
 'use strict';
 
-import { Position, FoldingRange, FoldingRangeKind } from 'vscode-languageserver-types';
+import { Position, Range, TextDocument, FoldingRange, FoldingRangeKind } from 'vscode-languageserver-types';
 import { DockerfileParser } from 'dockerfile-ast';
 
 export class DockerFolding {
 
-    private createFoldingRange(foldingRangeLineFoldingOnly: boolean, startLine: number, endLine: number, startCharacter: number, endCharacter: number, kind: FoldingRangeKind): FoldingRange {
+    private createFoldingRange(foldingRangeLineFoldingOnly: boolean, startLine: number, endLine: number, startCharacter: number, endCharacter: number, kind?: FoldingRangeKind): FoldingRange {
         if (foldingRangeLineFoldingOnly) {
             return {
                 startLine,
@@ -20,19 +20,43 @@ export class DockerFolding {
         return FoldingRange.create(startLine, endLine, startCharacter, endCharacter, kind);
     }
 
+    private getLineLength(document: TextDocument, line: number): number {
+        let text = document.getText(Range.create(line, 0, line, Number.MAX_SAFE_INTEGER));
+        let length = text.length;
+        let char = text.charAt(length - 1);
+        while (char === '\r' || char === '\n') {
+            length--;
+            char = text.charAt(length - 1);
+        }
+        return length;
+    }
+
     public computeFoldingRanges(content: string, lineFoldingOnly: boolean, limit: number): FoldingRange[] {
         if (limit < 1) {
             return [];
         }
 
+        const ranges = [];
         const dockerfile = DockerfileParser.parse(content);
+        const document = TextDocument.create("", "", 0, content);
+        for (const instruction of dockerfile.getInstructions()) {
+            const range = instruction.getRange();
+            if (range.start.line !== range.end.line) {
+                const startLineLength = this.getLineLength(document, range.start.line);
+                const endLineLength = this.getLineLength(document, range.end.line);
+                ranges.push(this.createFoldingRange(lineFoldingOnly, range.start.line, range.end.line, startLineLength, endLineLength));
+                if (ranges.length === limit) {
+                    // return if we've reached the client's desired limit
+                    return ranges;
+                }
+            }
+        }
         const comments = dockerfile.getComments();
         if (comments.length < 2) {
             // no folding if zero or one comment
-            return []
+            return ranges;
         }
 
-        const ranges = [];
         let found = false;
         let startRange = comments[0].getRange();
         let end = Position.create(startRange.start.line + 1, startRange.start.character);